added computatFilters_2.py
The added file calculates the spectral sensitivities you want to simulate and then uses a constrained linear least squares solver to find the best-fit relative weights of the different filters in order to achieve these sensitivities. Output files include a .csv file with the best-fit coefficients as well as a plot showing the desired vs. actual sensitivities achievable.